requirements.txt
Django==4.2.7
psycopg[binary]>=3.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
gunicorn==21.2.0
python-dotenv==1.0.0
djangorestframework-gis==1.0
